code-main/requirements.txt
# Add needed packages for project
coverage>=6
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
pytest>=7
parameterized==0.9.0
pre-commit==3.8.0
kaggle==1.6.17
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2
pyspellchecker==0.8.1
datasets==3.0.1
huggingface-hub==0.25.1
textstat==0.7.4
wordcloud==1.9.3
nltk==3.9.1
scikit-learn==1.5.2
seaborn==0.13.2
openai==1.51.2
transformers==4.45.2
setfit==1.1.0
wandb==0.18.5
spacy==3.5.2
language_tool_python==2.8.1
gensim==4.3.3
numpy==1.24.3
kagglehub==0.3.3
ipywidgets==8.1.5
textblob==0.18.0.post0
syllapy==0.7.2
lexical-diversity==0.1.1
sentencepiece
huggingface_hub
transformers
torch
huggingface_hub
joblib
